Add files via upload
This activity is a simple responsive website navigation bar built with Bootstrap 5. It features a dark themed design with a brand logo, search bar and dropdown menu. The navbar also automatically collapses into a hamburger menu on mobile screens and it closes when tapping outside the menu for better user experience.